Add files via upload
This project is based on a data shipping sales of adventure works pvt.ltd dataset . It is designed on power BI for an Interactive dashboard. It emphasizes and aims at improving the overall shipping sales of the company and helps to regain the customer database for different regions around the world. The major aspects to the data coming from the company is to predict about the future sales and products to be designed for based on the demands.